Move model-free C# helpers to CSharpText
Move identifier grammar, keyword policy, expression-body recognition, and member text layout into the dependency-free CSharpText leaf. Rewire consumers and transfer the owning tests without changing behavior.


Copilot-Session: bc2613de-a721-4efd-8390-56db684eefd7
