+++ src/CSharpText/CSharpText.csproj
@@ -16,6 +16,8 @@
   <ItemGroup>
     <InternalsVisibleTo Include="CSharpText.Tests" />
     <InternalsVisibleTo Include="DotnetInspector.CSharpBodySlicer" />
+    <InternalsVisibleTo Include="ILInspector.CSharp" />
+    <InternalsVisibleTo Include="ILInspector.Metadata" />
   </ItemGroup>
 
 </Project>
