+++ src/ILInspector.CSharp/CSharpFormatter.cs
@@ -1,4 +1,5 @@
 using System.Text;
+using CSharpText;
 using ILInspector.Metadata;
 
 namespace ILInspector.CSharp;
