+++ src/ILInspector.Metadata/ApiSurface.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
-using ILInspector.CSharp;
+using CSharpText;
 using ILInspector.Findings;
 
 namespace ILInspector.Metadata;
