+++ src/ILInspector.Metadata/ApiSurfaceExtractor.cs
@@ -5,7 +5,6 @@ using System.Reflection.Metadata.Ecma335;
 using System.Reflection.PortableExecutable;
 using System.Text;
 using CSharpText;
-using ILInspector.CSharp;
 
 namespace ILInspector.Metadata;
