+++ src/ILInspector.Metadata/ILInspector.Metadata.csproj
@@ -28,10 +28,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <!-- CSharp owns this spelling policy. Metadata source-links the internal
-         classifier while its legacy API surface still produces C# strings. -->
-    <Compile Include="..\ILInspector.CSharp\CSharpKeywords.cs" Link="CSharpKeywords.cs" />
-    <Compile Include="..\ILInspector.CSharp\CSharpIdentifierCore.cs" Link="CSharpIdentifierCore.cs" />
     <ProjectReference Include="..\CSharpText\CSharpText.csproj" />
     <ProjectReference Include="..\ILInspector.Findings\ILInspector.Findings.csproj" />
     <ProjectReference Include="..\InertText\InertText.csproj" />
