+++ src/dotnet-inspect.Tests/HttpClientFactoryTests.cs
@@ -111,7 +111,7 @@ public class HttpClientFactoryTests : IDisposable
     public async Task EnableNetworkTrafficLogging_PrintsTrafficKindWithoutBlocking()
     {
         using var error = new StringWriter();
-        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(ILInspector.CSharp.CSharpIdentifier.ContainRenderedText, error))
+        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(CSharpText.CSharpIdentifier.ContainRenderedText, error))
         {
             using var client = new HttpClient(new NetworkTelemetryHandler(
                 new StubHttpMessageHandler(),
@@ -139,7 +139,7 @@ public class HttpClientFactoryTests : IDisposable
     {
         using var error = new StringWriter();
         var transport = new StubHttpMessageHandler();
-        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(ILInspector.CSharp.CSharpIdentifier.ContainRenderedText, error))
+        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(CSharpText.CSharpIdentifier.ContainRenderedText, error))
         using (var client = new HttpClient(new NetworkTelemetryHandler(
             transport,
             NetworkClientKinds.Shared)))
@@ -179,7 +179,7 @@ public class HttpClientFactoryTests : IDisposable
     {
         using var error = new StringWriter();
         var transport = new StubHttpMessageHandler();
-        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(ILInspector.CSharp.CSharpIdentifier.ContainRenderedText, error))
+        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(CSharpText.CSharpIdentifier.ContainRenderedText, error))
         using (var client = new HttpClient(new NetworkTelemetryHandler(
             transport,
             NetworkClientKinds.Shared)))
@@ -236,7 +236,7 @@ public class HttpClientFactoryTests : IDisposable
                 NetworkClientKinds.Shared);
         }
 
-        var mermaid = diagram.ToMermaid(ILInspector.CSharp.CSharpIdentifier.ContainRenderedText);
+        var mermaid = diagram.ToMermaid(CSharpText.CSharpIdentifier.ContainRenderedText);
 
         Assert.Contains("flowchart TD", mermaid);
         Assert.Contains("n0[\"dotnet-inspect\"]", mermaid);
@@ -291,7 +291,7 @@ public class HttpClientFactoryTests : IDisposable
         _ = DotnetInspector.Core.CoreCache.TryGet("symbol-misses", key, extension: "forbidden");
         _ = DotnetInspector.Core.CoreCache.TryGet("symbol-misses", key, extension: "miss");
 
-        var mermaid = diagram.ToMermaid(ILInspector.CSharp.CSharpIdentifier.ContainRenderedText);
+        var mermaid = diagram.ToMermaid(CSharpText.CSharpIdentifier.ContainRenderedText);
 
         Assert.Contains("cache store<br/>symbol-misses/forbidden", mermaid);
         Assert.Contains("cache miss<br/>symbol-misses/miss", mermaid);
@@ -312,7 +312,7 @@ public class HttpClientFactoryTests : IDisposable
         CacheTelemetry.Record(category, key, CacheAccessResult.Store);
         CacheTelemetry.Record(category, key, CacheAccessResult.Hit);
 
-        var mermaid = diagram.ToMermaid(ILInspector.CSharp.CSharpIdentifier.ContainRenderedText);
+        var mermaid = diagram.ToMermaid(CSharpText.CSharpIdentifier.ContainRenderedText);
         var label = $"{category} {key}";
 
         Assert.Equal(1, CountOccurrences(mermaid, $"cache hit<br/>{label}"));
@@ -325,7 +325,7 @@ public class HttpClientFactoryTests : IDisposable
         bool allowTrafficKind)
     {
         using var error = new StringWriter();
-        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(ILInspector.CSharp.CSharpIdentifier.ContainRenderedText, error))
+        using (DotnetInspector.Core.HttpClientFactory.EnableNetworkTrafficLogging(CSharpText.CSharpIdentifier.ContainRenderedText, error))
         {
             using var client = new HttpClient(new NetworkTelemetryHandler(
                 new StubHttpMessageHandler(),
