+++ src/dotnet-inspect/dotnet-inspect.csproj
@@ -76,6 +76,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Using Include="CSharpText" />
     <Using Include="ILInspector.SourceLink" />
     <ProjectReference Include="..\CSharpText\CSharpText.csproj" />
     <ProjectReference Include="..\ILInspector.Analysis\ILInspector.Analysis.csproj" />
