+++ tests/ILInspector.Metadata.Tests/CSharpKeywordsTests.cs
@@ -1,6 +1,5 @@
 using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
-using ILInspector.CSharp;
 
 namespace ILInspector.Metadata.Tests;
 
@@ -10,21 +9,6 @@ public sealed class CSharpKeywordsTests
     {
         "await", "file", "init", "record", "required", "scoped",
     };
-
-    [Theory]
-    [MemberData(nameof(DeclarationContextualKeywords))]
-    public void DeclarationPolicy_EscapesConservativeContextualSet(string identifier)
-        => Assert.True(CSharpKeywords.RequiresDeclarationEscape(identifier));
-
-    [Theory]
-    [InlineData("await", true)]
-    [InlineData("file", false)]
-    [InlineData("init", false)]
-    [InlineData("record", false)]
-    [InlineData("required", false)]
-    [InlineData("scoped", false)]
-    public void BodyPolicy_PreservesDeclarationOnlyContextualIdentifiers(string identifier, bool expected)
-        => Assert.Equal(expected, CSharpKeywords.RequiresBodyEscape(identifier));
 
     [Fact]
     public void ApiSurfaceAndDeclarationQuery_UseDeclarationPolicyForContextualNames()
