+++ tools/DecompilerHarness/corpus/two-row-authored-corpus.jsonl
@@ -1,2 +1,2 @@
-{"assembly":"ILInspector.CSharp","assemblyVersion":"1.0.0.0","tfm":"release","type":"ILInspector.CSharp.CSharpDeclarationWriter","method":"RenderMemberUnit","overload":0,"signature":"`0(ApiType,ApiMember,CSharpDeclarationOptions,IReadOnlyList<string>)","metadataToken":100663442,"parameterCount":4,"ilSize":105,"sourceUrl":"https://raw.githubusercontent.com/richlander/dotnet-inspect/cdb4826be3812a851a72a2f4b6f4da6ee2449fad/src/ILInspector.CSharp/CSharpDeclarationWriter.cs","checksumAlgorithm":"SHA256","checksum":"F8EB70D17DB02B261CC738AF4FFA650DFC307AB5D6FCF0732BF09D51F54E1C27","authoredBody":"options ??= new CSharpDeclarationOptions();\n\n    var references = CollectMemberTypeReferences(member);\n\n    var plan = TypeNamePlan.Create(references, options);\n\n    var declaration = RenderMemberDeclarationCore(type, member, options, methodParameters);\n\n    declaration = plan.Apply(declaration);\n\n\n    if (options.TerminateMemberDeclaration && NeedsTerminator(declaration))\n        declaration += \";\";\n\n\n    var source = ComposeUnit([declaration], plan.GeneratedUsings, options);\n\n    return new CSharpRenderedDeclaration(source, plan.GeneratedUsings, plan.Diagnostics);"}
-{"assembly":"ILInspector.CSharp","assemblyVersion":"1.0.0.0","tfm":"release","type":"ILInspector.CSharp.CSharpExpressionBody","method":"FromSingleStatement","overload":0,"signature":"`0(string)","metadataToken":100663529,"parameterCount":1,"ilSize":179,"sourceUrl":"https://raw.githubusercontent.com/richlander/dotnet-inspect/cdb4826be3812a851a72a2f4b6f4da6ee2449fad/src/ILInspector.CSharp/CSharpExpressionBody.cs","checksumAlgorithm":"SHA256","checksum":"DC1A1AF45A35F48303896A50FB575B679559FE4DFED977498792E1B584FBEEEE","authoredBody":"var line = body.Trim();\n\n    if (line.Length == 0\n        || line.Contains('\\n')\n        || !line.EndsWith(';')\n        || line.StartsWith(\"//\", StringComparison.Ordinal)\n        || line.StartsWith(\"/*\", StringComparison.Ordinal))\n        return null;\n\n\n    line = line[..^1].TrimEnd();\n\n    if (line.StartsWith(\"return \", StringComparison.Ordinal))\n    {\n        var expression = line[\"return \".Length..].TrimStart();\n        return expression.Length == 0 ? null : expression;\n    }\n\n    if (line is \"return\")\n        return null;\n\n    if (line.StartsWith(\"throw \", StringComparison.Ordinal))\n        return line;\n\n    return IsStatementExpression(line) ? line : null;"}
+{"assembly":"ILInspector.CSharp","assemblyVersion":"1.0.0.0","tfm":"release","type":"ILInspector.CSharp.CSharpDeclarationWriter","method":"RenderMemberUnit","overload":0,"signature":"\u00600(ApiType,ApiMember,CSharpDeclarationOptions,IReadOnlyList\u003Cstring\u003E)","metadataToken":100663442,"parameterCount":4,"ilSize":105,"sourceUrl":"https://raw.githubusercontent.com/richlander/dotnet-inspect/3818808646ce072da59faee897f83f1966061a44/src/ILInspector.CSharp/CSharpDeclarationWriter.cs","checksumAlgorithm":"SHA256","checksum":"97A66F7359EB8307DEBCCC7F0BD0D9B6E3D496206758EF4A061113F9675F27E2","authoredBody":"options ??= new CSharpDeclarationOptions();\n\n    var references = CollectMemberTypeReferences(member);\n\n    var plan = TypeNamePlan.Create(references, options);\n\n    var declaration = RenderMemberDeclarationCore(type, member, options, methodParameters);\n\n    declaration = plan.Apply(declaration);\n\n\n    if (options.TerminateMemberDeclaration \u0026\u0026 NeedsTerminator(declaration))\n        declaration \u002B= \u0022;\u0022;\n\n\n    var source = ComposeUnit([declaration], plan.GeneratedUsings, options);\n\n    return new CSharpRenderedDeclaration(source, plan.GeneratedUsings, plan.Diagnostics);"}
+{"assembly":"ILInspector.CSharp","assemblyVersion":"1.0.0.0","tfm":"release","type":"ILInspector.CSharp.CSharpFormatter","method":"FormatMember","overload":0,"signature":"\u00600(ApiType,ApiMember,IReadOnlyList\u003Cstring\u003E)","metadataToken":100663583,"parameterCount":3,"ilSize":37,"sourceUrl":"https://raw.githubusercontent.com/richlander/dotnet-inspect/3818808646ce072da59faee897f83f1966061a44/src/ILInspector.CSharp/CSharpFormatter.cs","checksumAlgorithm":"SHA256","checksum":"1766A183E738F46A36DB8DF4A4502B23B9AEADCB3734638A5779CB8878829E84","authoredBody":"ArgumentNullException.ThrowIfNull(type);\n\n    ArgumentNullException.ThrowIfNull(member);\n\n    return CSharpDeclarationWriter.RenderMemberDeclaration(\n        type,\n        member,\n        _declarationOptions,\n        methodParameters);"}
